Merged PR 1703009: Fixed targets parsing
Previously targets starting with an ampersand (&) were parsed wrongly because the regex used to identify a target didn't account for ampersands. Also, because the regex didn't have a start anchor (^), we weren't getting a "failed to parse targets" error. Fixed the regex to allow ampersands, and anchored it so we get an error if other targets fail to be fully parsed.

Related work items: #3693919