Merged PR 1700032: Initial SecRule engine request scanning
Initial stab at the scanning engine. It does the basics. More thought out error handling will follow in a later PR. Also, capturing the actual content proved tricky (getting the position of the last char in the match is all Hyperscan returns by default), so something more elaborate for that will follow in a PR later too.

The changes under the secrule folder are fully agnostic of the Regex engine, so we could swap Hyperscan it for RE2::Set if we want for example. It also helps isolate our code from the Hyperscan, so we can unit test it in isolation. This is what I've done in reqscanning_test.go. It contains "pre-canned" answers for what Hyperscan would return, rather than calling the actual Hyperscan. In the hyperscan package folder there is an integration test that does something similar, but with the actual real Hyperscan.

Regarding the Dockerfile change: Just the Hyperscan version changed. This version puts the header files in /usr/include instead of /usr/include/hs. Want this because it's too awkward to add an extra include path to the Go module that does the C language bindings for Hyperscan.

Related work items: #3693919