Merged PR 1833385: Request body streaming in Azwaf, server side
Implemented request body streaming in Azwaf. This is the server side part. The client side part is in a separate PR in the AppGw repo.

In the protobuf, the input to the RPC is now marked stream. The input messages are therefore set to be either the first message, which is the one containing all header info, or else the subsequent messages, which is the rest of the request body. I couldn't figure out a way to enforce in the .proto file that the first message is always headers, and the following is always body chunks, so I just raise errors in the server if this protocol is not followed.

The body streaming is wrapped in a way where it's exposed simply as an io.Reader, which is what the JSON/XML/multipart parsers will be taking as input. This avoids ever having to store the entire request in mem at once.

Also checking in a test client I was using during dev, as I think it might be useful when troubleshooting in the future.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/1833385

This is the corresponding client side Nginx change:
https://msazure.visualstudio.com/One/_git/Networking-AppGW/pullrequest/1833411?_a=overview

Related work items: #4894119