Merged PR 1856146: Refactor actions
Here's a refactoring suggestion. In my opinion this is simpler, because it decouples parsing and execution-logic for actions. The parser this way is isolated to just produces a data-only abstract syntax tree, just describing the statements in a tree of objects, without any kind of "handlers". All execution logic is isolated to ruleEvaluator.go.

The old actionHandler and actionResult interface didn't really make sense for most actions in my opinion (it's only 2-3 actions that are disruptive and produce status codes), so I removed it in favor of a generic Action interface. This also nicely replaces the old RawActions field. So now the Action []Action field contains what you would expect: actions of the different types that exist. There are some actions we don't care so much about yet, so I'm leaving them as type RawAction.

Turned out it's more convenient to keep most things as Action's of different specific types, rather than parsing them into fields on the Rule and ActionStmt struct. Made it easier to share logic between SecRule and SecAction, and deferred the decision about precedence to the evaluator rather than the parser (like if there's a nolog action followed by a log action). This is why I got rid of the Nolog and Msg fields. While I was dealing with the msg field I also did a bit more plumbing for logging to deal with the two separate log message fields there are (msg and logdata).