Merged PR 1964916: Not shared HS scratch databases
Not shared Hyperscan scratch databases.

enginefactory.go has a coordination go-routine, which deals with reusing scratch spaces between requests. It interfaces with the engine via two channels: one that the engine can read from if it wants a set of scratch spaces, and one that it writes to when it's done with a set of scratch spaces.

I'm a little sad that the MultiRegexEngine interface got this extra complexity now. It was pretty elegant and non-hyperscan-specific previously. But I haven't been able to think of a cleaner model. I'm open to ideas though!