Merged PR 2001391: Streaming URL-decoder for body parsing
Streaming URL-decoder for body parsing, so we no longer need to have whole application/x-www-form-urlencoded request bodies in memory at once (that was a shortcut I took earlier).

Also fixed the mockReader to be more accurate, as I needed the additional accuracy for the tests I was doing on urlDecoder.

Additionally added some tests to the URL-decoder that sits over in reqscanner.go. That one is simpler because it doesn't need to operate on a stream, because the whole URL is in memory at once.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2001391