Merged PR 2057459: Use the new waf.Decision in Azwaf-in-Nginx-proc and enable Azwaf-in-Nginx-proc during normal builds
Use the new waf.Decision in Nginx-in-proc.

The reason went under the radar is that the in-proc version doesn't compile by default in the official build, because we don't have the Nginx header files available here. Made a mini mock version of the Nginx headers so we are at least able to compile this code and catch problems like this in the future. The resulting .so compiled with these mock headers won't actually work though, so used Go build tags to compile differently when building the real Nginx-in-proc .so file.