Merged PR 2120240: Part 1 of regex selectors: parse targets further
This parses the targets in targets in ruleparsing.go a bit further. Previously the target was stored as an unparsed modsecurity target string such as
```
&ARG:/ab.c/
```
This mean that essentially its parsing was spread around the places it was used.

Now instead its represented as
```
Target {
    Name: "ARG",
    Selector: "ab.c",
    IsCount: true,
    IsRegexSelector: true,
}
```

I actually also wanted to get rid of `Name` as a string, and use an enum instead, but decided to keep that out of this PR to not make it too large.

The motivation for this is to be able to let `scanTarget` take not just a single `targetName` param, but also a `targetPartName` param which will be matched against the selector (which for example can be a regex). It will then look up in a structure where the regex selectors are taken into account, and store in `rxMatches` using a Target object rather than a string. This is all too messy to do with the targets in an un-parsed form.

Related work items: #5411030