Merged PR 2176081: Validateurlencoding operator
Implemented the `@validateurlencoding` operator. For this operator to make sense in the secrules, I also had to make the request scanner not fail when it itself stumbles upon invalid URL encoded strings. For this purpose I have implemented a WeakURLUnescape, and let the URL-parsing using this function instead of Go's builtin, which would fail with an error if it meets something that isn't valid URL escaped.

This change, in combination with a change to support `REQUEST_BODY` that I am sending separately, makes the CRS test failures go from 101 to 90.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2176081