Merged PR 2180672: Count-targets
Implementing correct evaluation of count targets in the SecRule language, such as `&ARGS:hello`, which means the number of times there was an ARG with the name `hello`. Or even just `&ARGS`, which means the number of ARGS in total.

Re-purposed the targetsPresent map I had previously introduced, so it not only tracks whether a target was present, but also how many times it was present.

This makes 5 more of the CRS tests pass.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2180672