Merged PR 2176249: REQUEST_BODY target for SecRule engine
This was a painful one. It's not elegant at all...

It's needed in order to correctly handle SecRule's REQUEST_BODY target, which needs the entire raw request body, but only when the content-type is application/x-www-form-urlencoded. In order to not let this get out of control, I've made an upper limit on how big request bodies we support in this mode. I think it's reasonable with such a limit, because I think the use case for application/x-www-form-urlencoded will be just small form posts. Larger bodies with file uploads or XML or JSON payload will have other content types.

I think the error message I've written for when the limit exceeds describes the situation pretty ok: "Request body length exceeded the limit (%d bytes) while the WAF was scanning the entire request body as a single field. The OWASP Core Rule Set and possibly other SecRule-based rule sets require this scan when the request body content-type is set to application/x-www-form-urlencoded."

This change lets the body parser get a callback function it can call to decide whether to store and scan the entire request body as a single field. Did it via a callback in order to avoid to code this secrule-specific oddity into the body parser.

I'm open to ideas on how to make this more elegant.

This change, in combination with a change to support the `@validateurlencoding` operator that I am sending separately, makes the CRS test failures go from 101 to 90.

Will rebase to master branch once feature/allanbo/validateurlencoding2 is in.