Merged PR 2193841: Timestamps in logs and other logging fixes, part 1
This adds timestamps to logs. There are still a bunch of things that aren't right with logs. Put TODO's in this PR for them, and currently still working on a PR to restructure results-logging a bit, which should simplify some things that are awkward now, and make it pretty straightforward to show the action as "Detected" in detection-mode.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2193841