Merged PR 2196973: Per-request results logger
Changing the architecture of the results logger a bit. Letting it be per-request, instead of per-server. The part that synchronizes log file writing is still per-server though (now called `logFileWriter`). Letting ResultsLogger be per-request is convenient, because it lets us give all meta data information to the logger beforehand (such as transaction ID, isDetectionMode, URI, remote IP, etc.), so we don't have to pass all this information all the way through all the engines that otherwise don't care about this data. For example it simplifies the interface to `IPReputationEngine`, as it now no longer needs parts of the HTTPRequest object that it doesn't care about (previously just in order to be able to pass it to the results logger).

I decided to do this instead of continuing the anti pattern, in order to plumb through isDetectionMode for custom rules.

To easiest understand this change, I recommend you start by looking at the change to `/waf/server.go`. The rest kind of follows from this.

Manually tested too with the `grpcurl` commands in the README.md.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2196973