Merged PR 2246604: Capture-action support
Capture action-support. I let the multi regex engine store capture groups in its results struct, so we can populate these when the capture action is run.

In `evalPredicate` I made the distinction clearer between looking up a result from the req scan results, and late scanning (as in, scanning variable values during eval phase).

Implemented a regex operator that the evaluation engine uses for late scanning.

I noticed that the compareActions function wasn't doing anything, because it was taking its strings.Builder as a value-type and not a pointer, so the strings.Builder that got passed in would not be modified outside the function. This is the reason for a few more fixes in the ruleparsing_test.go file.

I'm struggling to resist refactoring things a bit, to use token lists instead of varMacroMatches/valMacroMatches. Postponed for now to get more tests written and passing first, so the refactoring will be safer.

CRS 3.0 FTW before:
```
    crs_test.go:189: Total tests: 444, Skip: 68, Pass: 302, Fail: 74
    crs_test.go:190: Total tests pass percent: 80%
```
after:
```
    crs_test.go:189: Total tests: 444, Skip: 68, Pass: 311, Fail: 65
    crs_test.go:190: Total tests pass percent: 82%
```

Related work items: #4951675, #5719996