Merged PR 2305804: Strongly typed target names
Got annoyed with having to do ToLower and []byte to string and string to []byte conversions on the target names (like ARGS, MATCHED_VARS_NAMES, etc.). Therefore decided to turn it into a strongly typed enum-like value. Unfortunately it didn't give any measurable performance improvement, but I still think it's worth it for making things a bit more type safe, and to avoid the risk of forgetting a ToLower somewhere.