Merged PR 2306136: Support for macro as val with with special case targets
Normally we do not support scanning for a macro in a target that is only available during the scanning phase. This is because we aim at only doing one scan through the request, and the macros are not yet populated at that point in time.

This PR adds support for some special cases of this. It allows macro as val with with targets REQUEST_LINE, REQUEST_METHOD, REQUEST_PROTOCOL, REQUEST_HEADERS:Host. This is because there are a few CRS rules that require this. Example, where the REQUEST_METHOD is a field that was part of the request and normally only available during the scan-phase (as opposed to for example something like `TX:myvar`):
```
...
SecAction \
 "id:900200,\
  phase:1,\
  nolog,\
  pass,\
  t:none,\
  setvar:'tx.allowed_methods=GET HEAD POST OPTIONS'"

...

SecRule REQUEST_METHOD "!@within %{tx.allowed_methods}" \
    "id:911100,\
    ....
```

This brings the CRS tests down by 4 failures (both for 3.0 and 3.1).

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2306136