Merged PR 2306033: Refactor ruleevaluator.go for readability
I think this makes processPhase() a bit more easy on the eyes.

Basically what I did is created runActions and runActionsAfterAllRuleItemsTriggered as methods instead of lambda funcs, and put them on a struct that they can use to alter the state on instead of altering the state of the outer vars as closures previously.

Creating a phaseEvaluation struct to keep state is also useful in preparation for a upcoming change we need to make: to run phase 1 before scanning the body for in case the ctl:forceRequestBodyVariable action happens during phase 1.