Merged PR 2334807: ForceRequestBodyVariable part 1: refactor rule evaluation
This pulls out the loop that loops over each of the five phases, out of the secrule package.

Also eliminated the separate `disruptivePhase` concept, since that is implied by the `waf.Decision` value. Also removed `statusCode`, since we weren't returning it all the way up and using it anyway, and CRS doesn't use it.

This is needed for forceRequestBodyVariable to work. Rule evaluation phase 1 has to run before scanning the body. When it has run, we will know if forceRequestBodyVariable is set to on. If it is set to on, then we need to scan the request body (this will come in pull request number two). Previously all 5 phases ran after the request body was already scanned.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2334807

Related work items: #5859094