Merged PR 2351249: URL-decode REQUEST_FILENAME and REQUEST_BASENAME
URL-decode REQUEST_FILENAME and REQUEST_BASENAME.

My previous understanding that these should be raw and not url-decoded was wrong. Tested with modsec. The reference manual says something about transformations and anti evasion for this field, so I think I misinterpreted it as it being raw, when it's actually not.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2351249