Merged PR 2355020: CRS test runner support for encoded_request
In the CRS FTW tests, some of the test requests are defined via a field called `encoded_request`. This field contains the base64 encoded full raw request. Since we are not working at the level of raw requests, but instead at the level of somewhat parsed requests, we need to do some parsing in the CRS test runner for this.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2355020