Merged PR 2380079: Put phase loop back into secrule package
I had previously taken the loop over the five phases out of the secrule package, and put it in the waf package. This was because we needed the secrule engine to run all phase 1 rules before starting the body parsing. Jeremie suggested to make some explicit functions for this rather than relying on things outside the waf package to know which phases to call.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2380079