Merged PR 2489016: Avoid embedded anchors in CRS rule 942330
Rule 942330 has an embedded start anchor. Specifically this part:
```
(?:^[\"'`\\\\]*?(?:[\d\"'`]+|[^\"'`]+[\"'`]))+
```
has a `^` inside a `+`-group. Hyperscan does not support this.

It looks like the purpose of this group is to find lines that begin with quotes or digits. The `+` makes it consume multiple of such lines if there are multiple. But even without the `+`, we would always find one if one is present.

To get rid of the embedded start anchor, I suggest removing the `+` from behind the group. It's only effect was to allow multiple matches of the group, but in this case it's enough to just match the last instance of the group for this rule to trigger. We can then also remove the `(?:)` as there is no longer any purpose of having this expression be a group. The part in question then becomes:
```
^[\"'`\\\\]*?(?:[\d\"'`]+|[^\"'`]+[\"'`])
```

The rule will trigger on the same input, but the matched data is now a little bit less. `%{TX.0}` will now only contain the last instance of the group above. I think this is not a problem. Especially since the logdata action also contains `%{MATCHED_VAR}`, which contains the entire variable that triggered this matched (not just the matched portion).

All the above also applies to the other similar part
```
(?:^[\"'`\\\\]*?[^\"'`]+[\"'`])+
```

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2489016

Related work items: #4812232