Merged PR 2757640: Make bodyparser shortcircuit evalrequest decision flow  [Part 1]
**Summary:**
Make bodyparser shortcircuit EvalRequest decision flow if there's an error. The integration of bodyparsing wasn't being tested because we never called `return` inside the bodyparsing block of EvalRequest. Not calling return means that the EvalRequest execution isn't interrupted. This means that if we get an Allow evaluation from other parts of the EvalRequest, it'll replace the Block that could've originated from BodyParsing.

By integrating it, I ran into some issues when doing manual end to end testing, which is the motivation behind these changes. The issue can be seen in my changes made to `lengthlimits.go`. The length comparison happened prior to counting the bytes in the buffer. This in turn caused the lengthlimits evaluation to be offset by 2048 bytes, the size of the buffer.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/2757640

Related work items: #6485699